feat: final cleanup
